fix: theme 패키지 파일구조 변경
