+++ packages/core/theme/src/utils/index.ts
@@ -1,5 +1,3 @@
 export { convertNestedValue, mapValueToKeys } from './normalize';
-export { compoundVariants } from './variants';
 export { systemProperties } from './systemProperties';
-
 export type { SystemProperties } from './systemProperties';
